feat(Participant): 중복되는 부분 추상클래스로 분리
